# packages/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Keeper Financial API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/recruitment/postings": {
      "get": {
        "tags": [
          "recruitment"
        ],
        "summary": "List Public Postings",
        "operationId": "list_public_postings_api_v1_recruitment_postings_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 25,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicPostingList"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/recruitment/postings/{slug}/applications/start": {
      "post": {
        "tags": [
          "recruitment"
        ],
        "summary": "Start Candidate Application",
        "operationId": "start_candidate_application_api_v1_recruitment_postings__slug__applications_start_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          },
          {
            "name": "x-dev-auth-sub",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dev-Auth-Sub"
            }
          },
          {
            "name": "x-dev-auth-email",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dev-Auth-Email"
            }
          },
          {
            "name": "x-dev-auth-verified",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "false",
              "title": "X-Dev-Auth-Verified"
            }
          },
          {
            "name": "x-dev-auth-aal",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "aal1",
              "title": "X-Dev-Auth-Aal"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CandidateApplicationResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Verified identity required"
          },
          "404": {
            "description": "Posting unavailable"
          },
          "409": {
            "description": "Identity conflict"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/recruitment/postings/{slug}": {
      "get": {
        "tags": [
          "recruitment"
        ],
        "summary": "Get Public Posting",
        "operationId": "get_public_posting_api_v1_recruitment_postings__slug__get",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicPosting"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/recruitment-postings": {
      "get": {
        "tags": [
          "recruitment"
        ],
        "summary": "List Admin Postings",
        "operationId": "list_admin_postings_api_v1_admin_recruitment_postings_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 25,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "x-dev-auth-sub",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dev-Auth-Sub"
            }
          },
          {
            "name": "x-dev-auth-aal",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "aal1",
              "title": "X-Dev-Auth-Aal"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminPostingList"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Admin denied"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "recruitment"
        ],
        "summary": "Create Admin Posting",
        "operationId": "create_admin_posting_api_v1_admin_recruitment_postings_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "x-dev-auth-sub",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dev-Auth-Sub"
            }
          },
          {
            "name": "x-dev-auth-aal",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "aal1",
              "title": "X-Dev-Auth-Aal"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PostingCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminPosting"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Admin denied"
          },
          "409": {
            "description": "Conflict"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/recruitment-postings/{posting_id}": {
      "patch": {
        "tags": [
          "recruitment"
        ],
        "summary": "Patch Admin Posting",
        "operationId": "patch_admin_posting_api_v1_admin_recruitment_postings__posting_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "posting_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Posting Id"
            }
          },
          {
            "name": "x-dev-auth-sub",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dev-Auth-Sub"
            }
          },
          {
            "name": "x-dev-auth-aal",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "aal1",
              "title": "X-Dev-Auth-Aal"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PostingUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminPosting"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Admin denied"
          },
          "404": {
            "description": "Posting not found"
          },
          "409": {
            "description": "Posting lifecycle or slug conflict"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/recruitment-postings/{posting_id}/{action}": {
      "post": {
        "tags": [
          "recruitment"
        ],
        "summary": "Transition Admin Posting",
        "operationId": "transition_admin_posting_api_v1_admin_recruitment_postings__posting_id___action__post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "posting_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Posting Id"
            }
          },
          {
            "name": "action",
            "in": "path",
            "required": true,
            "schema": {
              "enum": [
                "publish",
                "close",
                "archive"
              ],
              "type": "string",
              "title": "Action"
            }
          },
          {
            "name": "x-dev-auth-sub",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dev-Auth-Sub"
            }
          },
          {
            "name": "x-dev-auth-aal",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "aal1",
              "title": "X-Dev-Auth-Aal"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminPosting"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Admin denied"
          },
          "404": {
            "description": "Posting not found"
          },
          "409": {
            "description": "Posting lifecycle conflict"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AdminPostingList": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AdminPosting"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "AdminPostingList"
      },
...
      "CandidateApplicationResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "recruitment_posting_id": {
            "type": "string",
            "format": "uuid",
            "title": "Recruitment Posting Id"
          },
          "source_posting_slug": {
            "type": "string",
            "title": "Source Posting Slug"
          },
          "source_posting_title": {
            "type": "string",
            "title": "Source Posting Title"
          },
          "source_posting_version": {
            "type": "integer",
            "title": "Source Posting Version"
          },
          "schema_version": {
            "type": "string",
            "title": "Schema Version"
          },
          "revision": {
            "type": "integer",
            "title": "Revision"
          },
          "state": {
            "type": "string",
            "enum": [
              "draft",
              "submitted",
              "withdrawn"
            ],
            "title": "State"
          },
          "status": {
            "type": "string",
            "enum": [
              "application_started",
              "application_submitted",
              "withdrawn",
              "declined"
            ],
            "title": "Status"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "given_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Given Name"
          },
          "family_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Family Name"
          },
          "preferred_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Preferred Name"
          },
          "phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          },
          "city": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "City"
          },
          "region": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Region"
          },
          "country_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Country Code"
          },
          "preferred_contact_method": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Preferred Contact Method"
          },
          "available_from": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Available From"
          },
          "referral_source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Referral Source"
          },
          "referral_detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Referral Detail"
          },
          "interest_statement": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Interest Statement"
          },
          "relevant_experience": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Relevant Experience"
          },
          "employment": {
            "items": {
              "$ref": "#/components/schemas/EmploymentEntryResponse"
            },
            "type": "array",
            "title": "Employment"
          },
          "education": {
            "items": {
              "$ref": "#/components/schemas/EducationEntryResponse"
            },
            "type": "array",
            "title": "Education"
          },
          "privacy_acknowledged": {
            "type": "boolean",
            "title": "Privacy Acknowledged"
          },
          "information_accuracy_confirmed": {
            "type": "boolean",
            "title": "Information Accuracy Confirmed"
          },
          "privacy_disclosure_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Privacy Disclosure Version"
          },
          "privacy_acknowledged_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Privacy Acknowledged At"
          },
          "submitted_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Submitted At"
          },
          "withdrawn_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Withdrawn At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "recruitment_posting_id",
          "source_posting_slug",
          "source_posting_title",
          "source_posting_version",
          "schema_version",
          "revision",
          "state",
          "status",
          "email",
          "given_name",
          "family_name",
          "preferred_name",
          "phone",
          "city",
          "region",
          "country_code",
          "preferred_contact_method",
          "available_from",
          "referral_source",
          "referral_detail",
          "interest_statement",
          "relevant_experience",
          "employment",
          "education",
          "privacy_acknowledged",
          "information_accuracy_confirmed",
          "privacy_disclosure_version",
          "privacy_acknowledged_at",
          "submitted_at",
          "withdrawn_at",
          "created_at",
          "updated_at"
        ],
        "title": "CandidateApplicationResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PostingCreate": {
        "properties": {
          "slug": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$",
            "title": "Slug"
          },
          "title": {
            "type": "string",
            "maxLength": 160,
            "minLength": 1,
            "title": "Title"
          },
          "summary": {
            "type": "string",
            "maxLength": 500,
            "minLength": 1,
            "title": "Summary"
          },
          "body": {
            "type": "string",
            "maxLength": 5000,
            "minLength": 1,
            "title": "Body"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "slug",
          "title",
          "summary",
          "body"
        ],
        "title": "PostingCreate"
      },
      "PostingUpdate": {
        "properties": {
          "slug": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100,
                "minLength": 1,
                "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Slug"
          },
          "title": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 160,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "summary": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          },
          "body": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 5000,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Body"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "PostingUpdate"
      },
      "PublicPosting": {
        "properties": {
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "summary": {
            "type": "string",
            "title": "Summary"
          },
          "body": {
            "type": "string",
            "title": "Body"
          }
        },
        "type": "object",
        "required": [
          "slug",
          "title",
          "summary",
          "body"
        ],
        "title": "PublicPosting"
      },
      "PublicPostingList": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/PublicPostingSummary"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "PublicPostingList"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}